requirements.txt
alembic==1.4.2
click==7.1.1
Flask==1.1.2
Flask-JWT-Extended==3.24.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.8.0
importlib-metadata==1.6.0
inflect==4.1.0
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
zipp==3.1.0
